+++ src/pages/shopping-cart/ShoppingCartPage.tsx
@@ -3,7 +3,7 @@ import { useNavigate } from "react-router-dom";
 import { getShoppingCartData } from "../../api/cart";
 import { Header } from "../../components/common";
 import { useAPIDataContext } from "../../context/APIDataProvider";
-import { useOrderListContext } from "../../context/OrderListProvider";
+import { useOrderListContext } from "../../domain/order/context/OrderListProvider";
 import CartLayout from "../../domain/cart/components/cart-layout/CartLayout";
 
 const ShoppingCartPage = () => {
